% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about.R
\name{aboutServer}
\alias{aboutServer}
\title{About Tab Server}
\usage{
aboutServer(id, shared_store, shared_rx)
}
\arguments{
\item{id}{Module ID}

\item{shared_store}{Shared reactive store for app data}

\item{shared_rx}{Shared reactive values for app state}
}
\value{
Shiny server module
}
\description{
About Tab Server
}
