% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_server.R
\name{genome_server_js}
\alias{genome_server_js}
\title{Return the tags$script() required by load_annotation_track().}
\usage{
genome_server_js()
}
\description{
Include this once in your Shiny UI, e.g. inside tags$head():
tags$head(genome_server_js())
}
