% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homeServer.R
\name{home_server}
\alias{home_server}
\title{Home Tab Server}
\usage{
home_server(id, shared_store, shared_rx)
}
\arguments{
\item{id}{Module ID}

\item{shared_store}{A reactiveValues object to share data across modules}

\item{shared_rx}{A list of reactive values for cross-module communication}
}
\description{
Home Tab Server
}
