% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvServer.R
\name{igv_server}
\alias{igv_server}
\title{IGV Tab Server}
\usage{
igv_server(id, shared_store, shared_rx)
}
\arguments{
\item{id}{Module ID}

\item{shared_store}{A reactiveValues object to share data across modules}

\item{shared_rx}{A list of reactive values for cross-module communication}
}
\description{
IGV Tab Server
}
