% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utility.R
\name{load_panel_app_data}
\alias{load_panel_app_data}
\title{Load PanelApp Data}
\usage{
load_panel_app_data(file = NULL)
}
\arguments{
\item{file}{Path to the PanelApp TSV file. If NULL, loads the default file from the package.}
}
\value{
A data frame containing PanelApp data.
}
\description{
Utility function to load PanelApp data from a TSV file.
}
