% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utility.R
\name{make_boundary_table}
\alias{make_boundary_table}
\title{DB Utility
#' Utility functions to load and manage database files.}
\usage{
make_boundary_table(
  x,
  round_base = 10,
  slice_pct = 10,
  include_numeric = TRUE,
  include_factor = TRUE,
  add_row_id = TRUE,
  verbose = TRUE,
  logger = function(...) cat(...)
)
}
\description{
DB Utility
#' Utility functions to load and manage database files.
}
