% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utility.R
\docType{data}
\name{nthreads}
\alias{nthreads}
\title{Load VEP Annotations Mapping}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
nthreads
}
\arguments{
\item{file}{Path to the VEP annotations TSV file. If NULL, loads the default file from the package.}
}
\value{
A named list where each key is a consequence and the value is a vector of terms.
}
\description{
Utility function to load VEP consequence-to-term mappings from a TSV file.
}
\keyword{datasets}
