% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzleapp-package.R
\docType{package}
\name{puzzleapp}
\alias{puzzleapp-package}
\alias{puzzleapp}
\title{puzzleapp: A Toy Shiny App}
\description{
This package contains a Shiny application bundled as an R package.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
