% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzlecore_process.R
\name{puzzlecore_allele_count}
\alias{puzzlecore_allele_count}
\title{Compute Allele Count for a Sample
This function computes the allele count for a given sample based on the inheritance model,
status, and sex.}
\usage{
puzzlecore_allele_count(inher, status, sex)
}
\arguments{
\item{inher}{The inheritance model (e.g., "Homozygous Recessive", "Dominant/ De Novo", "Compound Heterozygous", "X-Linked Recessive").}

\item{status}{The status of the sample ("affected", "unaffected", or "NA").}

\item{sex}{The sex of the sample ("male" or "female").}
}
\value{
A string representing the allele count for the sample.
}
\description{
Compute Allele Count for a Sample
This function computes the allele count for a given sample based on the inheritance model,
status, and sex.
}
\examples{
puzzlecore_allele_count("Homozygous Recessive", "affected", "male") # returns "2"
puzzlecore_allele_count("Dominant/ De Novo", "unaffected", "female") # returns "0"
}
