% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzlecore_process.R
\name{puzzlecore_load_phenotype_data}
\alias{puzzlecore_load_phenotype_data}
\title{Load Phenotype Data
This function loads phenotype-to-genes data from a TSV file.}
\usage{
puzzlecore_load_phenotype_data(file)
}
\arguments{
\item{file}{Path to the phenotype TSV file.}
}
\value{
A data frame containing phenotype-to-gene mappings.
}
\description{
Load Phenotype Data
This function loads phenotype-to-genes data from a TSV file.
}
\examples{
# phenotype_data <- puzzlecore_load_phenotype_data("path/to/phenotype_file.tsv")
}
