% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzlecore_process.R
\name{puzzlecore_load_vep_consequences}
\alias{puzzlecore_load_vep_consequences}
\title{Load VEP Consequences Data
This function loads VEP consequences data from a TSV file.}
\usage{
puzzlecore_load_vep_consequences(file)
}
\arguments{
\item{file}{Path to the VEP consequences TSV file.}
}
\value{
A data frame containing VEP consequences information.
}
\description{
Load VEP Consequences Data
This function loads VEP consequences data from a TSV file.
}
\examples{
# vep_consequences <- puzzlecore_load_vep_consequences("path/to/vep_consequences_file.tsv")
}
