% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzlecore_filter.R
\name{puzzlecore_variant_filter}
\alias{puzzlecore_variant_filter}
\title{Variant filtering function for SNVs and SVs}
\usage{
puzzlecore_variant_filter(
  snv_data,
  sv_data,
  snv_filters,
  sv_filters,
  pedigree,
  allele_tab,
  panel_app_genes,
  vep_consequences,
  phenotype_data
)
}
\arguments{
\item{snv_data}{Data table of SNV variants}

\item{sv_data}{Data table of SV variants}

\item{snv_filters}{List of SNV filter parameters}

\item{sv_filters}{List of SV filter parameters}

\item{pedigree}{Data table representing the pedigree information}

\item{allele_tab}{Data table of allele counts}

\item{panel_app_genes}{Data table of PanelApp genes}

\item{vep_consequences}{Data table of VEP consequences}

\item{phenotype_data}{Data table of phenotype information}
}
\value{
Filtered data table of variants after applying filters and compound heterozygous filtering
}
\description{
Variant filtering function for SNVs and SVs
}
