% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcplots.R
\name{qcPlotsServer}
\alias{qcPlotsServer}
\title{QC Plots Server}
\usage{
qcPlotsServer(id, shared_store, shared_rx)
}
\arguments{
\item{id}{Module id}

\item{shared_store}{Shared values list}

\item{shared_rx}{Shared reactive values list}
}
\value{
Shiny server logic
}
\description{
QC Plots Server
}
