% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawFilterServer.R
\name{rawFilterServer}
\alias{rawFilterServer}
\title{Raw Filter Tab Server}
\usage{
rawFilterServer(id, shared_store, shared_rx)
}
\arguments{
\item{id}{Module ID}

\item{shared_store}{A reactiveValues object to share data across modules}

\item{shared_rx}{A list of reactive values for cross-module communication}
}
\description{
Raw Filter Tab Server
}
