% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzlecore_preprocess.R
\name{read_and_normalise_vcf}
\alias{read_and_normalise_vcf}
\title{Read VCF and normalise functional annotations to CSQ}
\usage{
read_and_normalise_vcf(vcf_path)
}
\arguments{
\item{vcf_path}{Path to gzipped VCF (.vcf.gz)}
}
\value{
list(vcf_data, vcf_header)
}
\description{
Returns data unchanged if CSQ is already present. Otherwise converts ANN
(SnpEff) or BCSQ (BCFtools) to a pseudo-CSQ field using the mapping tables
in inst/extdata/preprocess/.
}
\keyword{internal}
