% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterServer.R
\name{selectFiltersServer}
\alias{selectFiltersServer}
\title{Filter Tab Server}
\usage{
selectFiltersServer(id, shared_store, shared_rx)
}
\arguments{
\item{id}{Module ID}
}
\description{
Filter Tab Server
}
