% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_session.R
\name{setup_app_logging}
\alias{setup_app_logging}
\title{Setup application logging}
\usage{
setup_app_logging(
  level = "info",
  logs_dir = log_default_dir(),
  older_than_days = 100,
  console = interactive()
)
}
\arguments{
\item{level}{Logging threshold: "info", "debug", etc.}

\item{logs_dir}{Directory to store logs.}

\item{older_than_days}{Purge logs older than this many days.}

\item{console}{Whether to log to console.}
}
\description{
Initializes per-app logging with lgr, purges old logs, optionally logs to console.
}
