% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzlecore_preprocess.R
\name{snv_format_fields}
\alias{snv_format_fields}
\title{Resolve FORMAT fields for SNV from the mapping, return selected field names.}
\usage{
snv_format_fields(mapping)
}
\description{
Resolve FORMAT fields for SNV from the mapping, return selected field names.
}
\keyword{internal}
