% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzlecore_preprocess.R
\name{sv_format_fields}
\alias{sv_format_fields}
\title{Resolve FORMAT fields for SV from the mapping, return selected field names.}
\usage{
sv_format_fields(mapping)
}
\description{
Resolve FORMAT fields for SV from the mapping, return selected field names.
}
\keyword{internal}
