% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dataServer}
\alias{dataServer}
\title{Data Module Server}
\usage{
dataServer(id, shared_store, shared_rx, dataset_names = NULL, prefix = "")
}
\arguments{
\item{id}{Module ID}

\item{shared_store}{Shared store (named list)}

\item{shared_rx}{Shared reactive values (list of reactiveVal)}

\item{dataset_names}{Optional character vector of allowed dataset names (NULL = all)}

\item{prefix}{Optional prefix for handling multiple different datasets}
}
\description{
Data Module Server
}
