% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_viewer.R
\name{log_viewer_ui}
\alias{log_viewer_ui}
\title{Log Viewer Module}
\usage{
log_viewer_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
Shiny UI object (log_viewer_ui) or NULL (log_viewer_server)
}
\description{
Log Viewer Module
}
