% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzlecore_preprocess.R
\name{process_sv_data}
\alias{process_sv_data}
\title{Process SV VCF to wide annotated table (legacy)}
\usage{
process_sv_data(
  svs_vcf,
  pedigree_data,
  svs_vcf_cohort = NA,
  svlog_static = NULL,
  svlog_db = NULL
)
}
\arguments{
\item{svs_vcf}{Path to gzipped SV VCF (.vcf.gz)}

\item{pedigree_data}{data.table with columns sample_id, kinship}

\item{svs_vcf_cohort}{Optional cohort VCF (.vcf.gz) for N_Cohort (default NA)}

\item{svlog_static}{Optional SVlog static TSV file for annotation (default NULL)}

\item{svlog_db}{Optional SVlog database TSV file for annotation (default NULL)}
}
\value{
data.table
}
\description{
Process SV VCF to wide annotated table (legacy)
}
