% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzleapp-package.R
\docType{package}
\name{puzzleapp-package}
\alias{puzzleapp}
\alias{puzzleapp-package}
\title{puzzleapp: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
