% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_PACKAGE.R
\docType{package}
\name{puzzleapp-package}
\alias{puzzleapp}
\alias{puzzleapp-package}
\title{PuzzleApp}
\description{
A Shiny App to visualise and analyse variants.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
