% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzleapp-package.R
\docType{package}
\name{puzzleapp-package}
\alias{puzzleapp}
\alias{puzzleapp-package}
\title{puzzleapp: Visualise and analyse variants}
\description{
an R/Shiny application for interactive exploration and review of genomic variants.
}
\author{
\strong{Maintainer}: Hiruna Samarkoon \email{hiruna.samarkoon@garvan.org.au}

Authors:
\itemize{
  \item Andre Martins Reis \email{a.martinsreis@garvan.org.au}
  \item Sasha Jenner \email{sjen6644@uni.sydney.edu.au}
}

}
\keyword{internal}
