% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzlecore_process.R
\name{puzzlecore_load_panel_app_data}
\alias{puzzlecore_load_panel_app_data}
\title{Load and Process PanelApp Data
This function loads PanelApp data from a specified file and processes it to extract relevant information.}
\usage{
puzzlecore_load_panel_app_data(file)
}
\arguments{
\item{file}{The path to the PanelApp data file.}
}
\value{
A data.table containing the processed PanelApp gene data.
}
\description{
Load and Process PanelApp Data
This function loads PanelApp data from a specified file and processes it to extract relevant information.
}
\examples{
# panel_app_genes <- puzzlecore_load_panel_app_data("path/to/panelapp_file.tsv")
# Processed behavior of the Sources column:
# Original Sources              -> Processed Sources
# ------------------------------------------------
# "Expert Review Green"         -> "Green"
# "Expert Review Red"           -> "Red"
# Other                      -> Unclassified"
}
