% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzlecore_process.R
\name{puzzlecore_parse_filter_table}
\alias{puzzlecore_parse_filter_table}
\title{Parse Filter Table
This function parses a filter table from a given file path and extracts SNV and SV filters.}
\usage{
puzzlecore_parse_filter_table(path)
}
\arguments{
\item{path}{The path to the filter table file or a data.table object.}
}
\value{
A list containing two elements: snv_filters and sv_filters, each being a list of filter parameters.
}
\description{
Parse Filter Table
This function parses a filter table from a given file path and extracts SNV and SV filters.
}
\examples{
# filters <- puzzlecore_parse_filter_table("path/to/filter_table.tsv")
}
